+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Added support for RabbitMQ 3.13, with Route, Partitions and SuperStreams support
 - Added support for `:route`, `:partitions`, `:consumerupdate`, `:exchangecommandversions`, `:streamstats`, `:createsuperstream` and `:deletesuperstream` commands.
 - Added serialization options for encoding and decoding messages.
 - TLS Support
+- Functional `single-active-consumer` and `filter_value` properties for consumers
 
 ### 0.4.0 Changes